Rename tests/testdata.py as tests/utils.py
