+++ tests/test_wps_generate_climos.py
@@ -1,7 +1,7 @@
 import pytest
 
 
-from .testdata import TESTDATA, process_err_test
+from .utils import TESTDATA, process_err_test
 from wps_tools.testing import run_wps_process, local_path, url_path
 from thunderbird.processes.wps_generate_climos import GenerateClimos
