+++ tests/test_wps_generate_prsn.py
@@ -2,7 +2,7 @@ import pytest
 import re
 
 
-from .testdata import TESTDATA
+from .utils import TESTDATA
 from wps_tools.testing import run_wps_process, local_path, url_path
 from thunderbird.processes.wps_generate_prsn import GeneratePrsn
